refactor: split app.rs and ui.rs into submodules
app.rs (1705 lines) -> app/ module:
- app/mod.rs: App struct and impl methods
- app/kv_editor.rs: KvEditorState for key-value editing
- app/clipboard.rs: system clipboard integration
- app/url_utils.rs: URL encode/decode and name extraction

ui.rs (1747 lines) -> ui/ module:
- ui/mod.rs: entry point, colors, layout
- ui/sidebar.rs: sidebar rendering
- ui/request.rs: request panel tabs (body, headers, auth, params)
- ui/response.rs: response panel with search highlighting
- ui/modals.rs: popups (method, history, curl, postman, env)
- ui/settings.rs: settings view
